+++ src/Controller/DeliveryNoteController.php
@@ -58,7 +58,12 @@ class DeliveryNoteController extends AbstractController implements PushInterface
 
         $trackingCodes = [];
         foreach ($deliveryNote->getTrackingLists() as $trackingList) {
-            $trackingCodes = \array_merge($trackingCodes, $trackingList->getCodes());
+            foreach ($trackingList->getCodes() as $code) {
+                foreach ($trackingList->getTrackingURLs() as $trackingUrl) {
+                    #if tracking url is sent, use url, otherwise use tracking code
+                    $trackingCodes[] = str_contains($trackingUrl, $code) ? $trackingUrl : $code;
+                }
+            }
         }
 
         if (!empty($trackingCodes)) {